Revert "Implement getblock (verbosity 1) RPC call (v29.2)"
This reverts commit 34830bffe35a013e8635795dd93ab580139b03b6.
